Skiller ut useFileUploadAllowed til egen fil
